pyproject.toml
[project]
name = "basic-rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.14",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.24",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.56.0",
    "torchvision>=0.26.0",
]
